requirements.txt
numpy>=1.19
pandas>=1.1
tf-models-official>=2.4.0,<=2.7.0
ml_collections>=0.1.0
absl-py>=0.13.0

pysam
